Separated out PadCast.cpp and PadCast.h, now using a main.cpp for main()
