+++ .gitignore
@@ -1,3 +1,6 @@
+# PadCast archive
+.archive/
+
 # Prerequisites
 *.d
